+++ bookingapp/src/main/java/com/dentalcura/bookingapp/dao/impl/PatientDAOH2.java
@@ -66,13 +66,13 @@ public class PatientDAOH2 implements IDao<Patient> {
             connection = DriverManager.getConnection(DB.URL,DB.USR,DB.PWD);
             preparedStatement = connection.prepareStatement(SQLQueries.PATIENT.getInsertCustom());
 
-            preparedStatement.setLong(1, patient.id());
+            preparedStatement.setInt(1, patient.id());
             preparedStatement.setString(2, patient.name());
             preparedStatement.setString(3, patient.surname());
             preparedStatement.setInt(4, patient.niNumber());
             preparedStatement.setString(5, patient.registrationDate());
             // fk address_id
-            preparedStatement.setLong(6, patient.address().id());
+            preparedStatement.setInt(6, patient.address().id());
 
             preparedStatement.executeUpdate();
             preparedStatement.close();
@@ -103,7 +103,7 @@ public class PatientDAOH2 implements IDao<Patient> {
             ResultSet resultSet = preparedStatement.executeQuery();
 
             while (resultSet.next()){
-                Long id = resultSet.getLong(1);
+                int id = resultSet.getInt(1);
                 String name = resultSet.getString(2);
                 String surname = resultSet.getString(3);
                 int niNumber = resultSet.getInt(4);
@@ -148,7 +148,7 @@ public class PatientDAOH2 implements IDao<Patient> {
     }
 
     @Override
-    public Patient selectByID(Long id) {
+    public Patient selectByID(int id) {
         Connection connection;
         PreparedStatement preparedStatement;
         Patient patient = null;
@@ -162,11 +162,11 @@ public class PatientDAOH2 implements IDao<Patient> {
             Class.forName(DB.DRIVER);
             connection = DriverManager.getConnection(DB.URL,DB.USR,DB.PWD);
             preparedStatement = connection.prepareStatement(SQLQueries.PATIENT.getSelectById());
-            preparedStatement.setLong(1, id);
+            preparedStatement.setInt(1, id);
             ResultSet resultSet = preparedStatement.executeQuery();
 
             while (resultSet.next()){
-                Long tabId = resultSet.getLong(1);
+                int tabId = resultSet.getInt(1);
                 String name = resultSet.getString(2);
                 String surname = resultSet.getString(3);
                 int niNumber = resultSet.getInt(4);
@@ -205,7 +205,7 @@ public class PatientDAOH2 implements IDao<Patient> {
             connection = DriverManager.getConnection(DB.URL,DB.USR,DB.PWD);
             preparedStatement = connection.prepareStatement(SQLQueries.PATIENT.getUpdateById());
 
-            preparedStatement.setLong(5, patient.id());
+            preparedStatement.setInt(5, patient.id());
 
             preparedStatement.setString(1, patient.name());
             preparedStatement.setString(2, patient.surname());
@@ -245,7 +245,7 @@ public class PatientDAOH2 implements IDao<Patient> {
     }
 
     @Override
-    public Patient deleteByID(Long id) {
+    public Patient deleteByID(int id) {
         Connection connection;
         PreparedStatement preparedStatement;
 //        Patient patient = selectByID(id);
@@ -256,7 +256,7 @@ public class PatientDAOH2 implements IDao<Patient> {
             connection = DriverManager.getConnection(DB.URL,DB.USR,DB.PWD);
             preparedStatement = connection.prepareStatement(SQLQueries.PATIENT.getDeleteById());
 
-            preparedStatement.setLong(1, id);
+            preparedStatement.setInt(1, id);
 
             preparedStatement.executeUpdate();
             preparedStatement.close();
