+++ bookingapp/src/main/java/com/dentalcura/bookingapp/model/Address.java
@@ -1,7 +1,7 @@
 package com.dentalcura.bookingapp.model;
 
 
-public record Address (Long id, String streetName, int streetNumber, int floor, String department){
+public record Address (int id, String streetName, int streetNumber, int floor, String department){
 
     @Override
     public String toString() {
